HeaderSlices: preverified headers downloader
The preverified headers are downloaded with a help of a HeaderSlices data structure.
HeaderSlices is a memory-limited buffer of HeaderSlice objects, each containing a sequential slice of headers.
Each HeaderSlice has a status.

Downloading happens with several stages where each of the stages processes blocks in one status, and updates them to proceed to the next status. All stages runs in parallel, although most of the time only one of the stages is actively running, while the others are waiting for the status updates or timeouts.

Stages:

* FetchRequestStage - sends requests to P2P via sentry to get the slices. Slices become Waiting.
* FetchReceiveStage - receives the slices, and sets Downloaded status.
* RetryStage - handles timeouts. If a slice is Waiting for too long, we need to request it again. Status is updated to Empty (the slice will be processed by the FetchRequestStage again).
* VerifyStage - checks that block hashes are matching the expected ones and sets Verified status.
* SaveStage - saves slices into the database, and sets Saved status.
* RefillStage - forgets the Saved slices from memory, and creates more Empty slices until we reach the end of preverified chain.