RetryStage: retry requests only if needed.
If a headers slice fetch request was sent, it gets Waiting status.
When it is Waiting for too long, it can timeout.

Apply the retry logic for each slice individually if it times out.
The timeout has a back off logic: it increases if there's no reply with each retry attempt.