Header downloader: SaveStage.
SaveStage is by saving the verified headers
into the "Header" table in the key-value database.
It uses the header hashes as keys and RLP data as values.

HeaderSlices is refactored so that find_by_status
is not holding the overall lock.
It returns a detached RC reference to a single HeaderSlice.
This allows to hold the HeaderSlice lock while blocking on await in save_pending,
and let other stages proceed at the same time.