Header downloader: queue optimizations.
* avoid retrying queued sent messages by having only 1 message in the channel at a time.
if it is > 1 there's a chance that it is not sent to congested upstream within 5 sec,
and after retry we have 2 identical requests in the queue. When they get executed
we are getting a "slice that we didn't request".

* avoid premature SentryClientReactorEventLoop exit causing fetch_request_stage_stream failure: Err(ReactorStopped)

* refactor SentryClientReactorEventLoop to be able to send and receive simultaneously
previously if sentry.send_message was blocked on await, the loop was making no progress
even if the incoming messages were in in_stream.next.
Now send and receive streams are processed in parallel.