Reconnect to sentry.
sentry_client_connector module implements the basic reconnection logic: connecting to sentry GRPC and retrying every 5 sec on failure. This logic is reused for the initial connection (before status is sent), and also for disconnects happening during downloading.

connector_stream is a stream that produces the initial connected sentry client, and then new clients after each reconnect. polling this stream for the first time returns the connected client. polling it for the second, third etc. time forces a new connection (after some retries) and a new client.

When a disconnect is detected (during send or receive), we have to reconnect and recreate the send/receive streams providing a new sentry client. During that time the send/receive handling is paused (by removing the streams from the stream map), and then restarted (by adding new streams).

When reactor is stopped externally, we need to stop reconnecting. In order to achieve this the sentry connector stream is attached to the same reactor loop that reacts to stop signals.

Note: mpsc channel is good for controlling back pressure, because it blocks or returns a failure when it is full, but it only supports one subscriber (send_message_receiver). We need to have an exclusive mutable reference in order to get messages from it. But for reconnecting we actually need to stop receiving and restart it with a different streams setup. In order to share this subscriber across reconnects a tokio::sync::Mutex is used. The initial send_stream holds this mutex exclusively. When a disconnect happens, the send stream is discarded and the mutex is unlocked. When we reconnect, the new send stream is attached and it retakes the mutex.