Downloader: verify_slice_difficulties.
During the linear header download phase we need to verify that the header difficulty field is correct
with respect to its parent header (and the current chain config rules).

VerifyStageLinear stage checks that this relation holds between each sequential pair of a headers slice.
VerifyStageLinearLink stage checks this relation on the slice edges - between the last header of a previous verified slice (last verified header) and the first header of an internally verified slice.