Downloader: fix save_pending_monotonic logic.
The logic was based on indexes, but they are not stable
since save_slice() is async, RefillStage might or might not trigger
in between and remove the saved slices.

Also it was possible to not have RefillStage call between SaveStage calls.
In this case we've got Saved status at front, and confused it with a non-monotonic case.

Fixed by searching for the next slice in each loop iteration,
skipping already processed ones.