Downloader: cache hash() v2 (inside downloader).
This caches the hash() to avoid the 2nd recomputation in SaveStage.

Previously it was computed twice:
* In VerifyStage to connect hash(parent) to child.parent_hash;
* In SaveStage to obtain a HeaderKey for saving.

With this caching it saves about 17% of header downloader running time:

This bumps the max speed from about 4-5K blocks/sec to about 5-6K blocks/sec,
which translates to about 10-15 minutes saving on mainnet.