Downloader: fix linear phase hang if saved to fast.
If a slice is verified and saved too fast,
the next slice monotonic verification aborts.
We can't abort on Saved slices, because they will be evicted,
and so the remaining_count won't change (potentially),
causing the VerifyStageLinearLink to wait forever.

It is fine to abort on Invalid, because they are gonna be redownloaded,
verified internally, and the remaining_count will change.