Downloader: fork mode start test.
Fork mode starts when a slice fails to link to the canonical chain
in the forky phase for the first time (in VerifyStageLinearLink).
Right when it starts, the fork and canonical chain continuation blocks
are updated for Refetch (ForkModeStage::setup).

The fork mode start test verifies that given this condition,
the VerifiedInternally slice that fails to link (denoted as "|=")
gets marked with a Fork status (denoted as "Y"),
and that it is backed up to fork_status and becomes emtpy
(denoted as "-/Y") for the canonical chain continuation.

In addition, the canonical chain last Saved slice (denoted as "+")
is backed up to fork_status and becomes emtpy (denoted as "-/+")
for the fork chain continuation.