Downloader: basic verify link test.
Mode::Linear of VerifyStageForkyLink continues for as long as
it is possible to link slices to the canonical chain.

The test verifies that given a VerifiedInternally slice ("=")
it transitions to Saved ("+").
This has to be different from a slice with a broken link ("|="),
which must start Mode::Fork (in the fork_mode_start() test).

To distinguish the situations a HeaderSliceVerifierMock is used.
HeaderSliceVerifier is changed to a trait in stages
in order to substitute the verification logic for tests.