Downloader: tests check IDs.
Generate fixed IDs instead of sequential IDs for the test headers.

By default the IDs match the block numbers.
Custom IDs are denoted as 'a, 'b, 'c etc.
For each slice headers a range of IDs is assigned:
'a = [0..192], 'b = [192..384], 'c = [384..576] etc.

HeaderSliceVerifierMock logic is simplified:
verify_link succeeds if the blocks have sequential IDs.
For example, 'c can be linked to 'd, but not to 'f,
while 'e can be linked to 'f.

This obsoletes the "|=" notation,
adds more checks for block positions in result/forked
and prepares for more complicated test cases.