Downloader: fork unwind.
When switch to fork is needed, exit the forky stage loop,
and make an unwind request for the main stage loop.

Process unwind() call by the DB cleanup.

When execution starts again, finalize the switch_to_fork() changes.
Note: in tests there's no unwind, so finalize is called immediately after run
to verify the slices state after unwind.

The stage termination and unwind request is based on the ForkSwitchCommand,
but can be based on some other command.
This can be useful if there's a different situation thata requires unwinding.