Fix panic in interhashes stage progress message
During the trie update, we log an estimate of the remaining time every
30 seconds, for which the elasped time is divided by a progress ratio.
For very large updates, like the initial sync of Ethereum mainnet,
the measured progress may still be zero when the first message is due.
We check for this in order to not divide by zero and panic when
computing the Duration.